requirements.txt
requests
python-dotenv
google-generativeai
openai==0.28
numpy
pandas